Organizes client container into folder and updates imports
